// src/app/shared/models/remision.model.ts
// Generated by https://quicktype.io

export interface Remision {
    id:                 number;
    correlativo?:      string;
    idCliente?:          number;
    codigoCliente?:      string;
    nombreCliente?:      string;
    idSucursal:         number;
    numeroSucursal?:         number;
    fecha?:              string;
    direccion?:              string;
    observacion?:              string;
    itemsEliminados?:    number[]|null;
    detalle?:            RemisionDetalle[];
    nit?:         number;
}

export interface RemisionDetalle {
    id?:              number;
    idRemision?:         number;
    correlativoVenta?:         String;
    codigoProducto?:  string;
    codigo?:     string;
    producto?:  string;
    cantidad:        number;
}